Move source to MTpy package folder
This moves source files into the right folder structure for a package named
MTpy, and subpackages MTpy.core, MTpy.imaging, and MTpy.utils.

Note that because the __init__.py files are empty, all modules need to be
imported explicitly for use:

    >>> import MTpy
    >>> dir(MTpy)
    ['__builtins__', '__doc__', '__file__', '__name__', '__package__', '__path__']
    >>> import MTpy.core
    >>> dir(MTpy.core)
    ['__builtins__', '__doc__', '__file__', '__name__', '__package__', '__path__']
    >>> import MTpy.core.OCCAMTools
    >>> dir(MTpy.core)
    ['OCCAMTools',
     '__builtins__',
     '__doc__',
     '__file__',
     '__name__',
     '__package__',
     '__path__']
    >>> import MTpy.core.WS3DTools
    >>> import MTpy.core.BIRRPTools
    >>> dir(MTpy.core)
    ['BIRRPTools',
     'MTtools',
     'OCCAMTools',
     'WS3DTools',
     '__builtins__',
     '__doc__',
     '__file__',
     '__name__',
     '__package__',
     '__path__']

Note that MTpy.core.MTtools was imported indirectly.

Future changes
--------------

We could change this import style in the future so that a select number
of functions and classes were imported into the MTpy namespace, by putting
something like the following in MTpy\__init__.py:

    __all__ = ['some_func', 'some_other_func']
    from MTpy.core.MTtools.some_func
    from MTpy.imaging.ptplots.some_other_func

Then the following would work:

    >>> import MTpy
    >>> dir(MTpy)
    ['some_func',
     'some_other_func',
     '__builtins__',
     '__doc__',
     '__file__',
     '__name__',
     '__package__',
     '__path__']

Or into the global namespace (although this style seems frowned-upon):

    >>> from MTpy import *
    >>> globals().keys()
    [...,
     'some_func',
     'some_other_func',
     ...]

This is the way that many projects are organized, like NumPy and matplotlib.
It has the advantage of keeping code well-organized, but making it accessible
without typing very long module names and paths.