Remove dash from wsmt-pv.py
The dash needs to be removed because the "entry_points" argument in the
setuptools setup function requires scripts to be importable Python
modules, and dashes are not allowed in such module names.

The advantages of using setuptools entry_points, when it's installed,
are such that I think we should forgo dashes in module names.