Add package setup script
Adds a script in the top directory to install MTpy as a Python
package.

Operation
---------

The script caters for two installation scenarios, but is made
so there is no duplication of hand-maintained information.

Scenario A: setuptools is installed. Then the script uses the
"entry_points" argument to setup(). This allows two things: creating
a console script which points to a function within a module, rather
than a whole script; and it wraps scripts in a EXE file on Windows,
making them discoverable by the shell on the PATH variable, making
Tab-completing their names possible.

Scenario B: only distutils is installed. Then only scripts are
available, and they are copied over to the system scripts folder.
Also, distutils does not automatically discover all sub-packages,
so they must be listed.

Maintenance
-----------

There are two parts in this file which need to be maintained by hand:

1. A list of console scripts. This is specified in the "entry_points"
   format, and automatically corrected for distutils if necessary.

2. A list of sub-packages (e.g. 'MTpy.core', 'MTpy.imaging').

Usage
-----

To install the package, if you have setuptools installed:

    $ python setup.py develop

This will install the code in your system site-packages as links to
the original source, meaning that changes will be reflected on reload()
commands.

Or otherwise:

    $ python setup.py install

will copy the source files over to site-packages.

Future changes
--------------

If we release MTpy on PyPi (and I think we should, once it's more
developed), we will need to put metadata arguments in the setup()
function call here, as outlined here:

http://docs.python.org/distutils/setupscript.html#additional-meta-data
http://docs.python.org/distutils/packageindex.html