Fix PhaseTensor Plots
A range of parameters can be passed to the PlotPhaseTensorMaps class,
which is a sub-class of two other classes above in the hierarchy.
Keyword arguments passed to this class as **kwargs were not being
handled and filtered consistently down the class chain.

We now process keyword arguments in each super class and filter them
down the class chain, such that at the end of
PlotPhaseTensorMaps.__init__(), we are only left with keywords arguments
external to the class, i.e. those that are to be passed to matplotlib
e.g. alpha, edgecolor, etc.