Fix test_modem_inputfiles_builder tests
The `data.project_stations_on_topography` function has been modified
to append 'topo' to written data filename. This meant the wrong
test output was being compared to the baseline output. The test
output to be compared has changed to 'ModEM_Datatopo.dat'.

The same function has also been changed to increase the precision of
the value added to 'topoval' when calcuting elevation - from 0.1
to 0.001. This required a new baseline file to be created to account
for increase in precision.

Also removed redundant call to `write_data_file` in test intended
to zero the elevation in data file - this is now done in
`write_data_file` itself if `elevation` param is False.