Use model attributes as default parameters
The model contains attributes for padding and cell resolution. These
were not being used in the conversion. Instead values were provided
by the user with hardcoded fallbacks. Change the fallback to be the
actual value contained by the model object, reducing the amount
of parameters that need to be entered.

Also factor out listing depths into its own function.