Fix bug when reading padding from model file
In read_model_file(), the padding is determined by counting the number
of nodes east and north that aren't equal to the east and north cell
sizes respectively.

The value is off by 1. The indicies of the cells that don't match,
i.e. padding cells, are put in an array and the last value of the
array is taken as the  number of padding cells. This should actually
be the size of the array, i.e. the total number of padding cells.
Because indicies are 0...N - 1, the last value is 1 less than the
number of padding cells.

In the case of the north padding, it also wasn't being assigned
correctly. It was being set to `model.north_pad` when the actual
attribute is `model.pad_north`.

If any calculations have deen done that rely on the padding values of
a model object post-reading of a model file, they need to be corrected.