Fix ordering of trimming pad cells on meshgrid
The trimming of padding cells off the X and Y arrays of the
resistivity grid were reversed (Y array being trimmed with X padding
and vice versa). This was causing the error when differing X and Y pad
values were used.

The error was a mismatch  between points (from mesh grid cells) and
values (from the res grid) to be interpolated, which caused the numpy
interpolation function to raise an error. The mesh grid was trimmed
correctly, leading to a mismatch in the number of elements with the
incorrectly trimmed res grid.