Make relative station elevation same as elevation
According to Data.project_stations_on_topography, relative station
elevation is relative to the model surface. So before topography is
applied, station elevation can only be relative to the spheroid of the
CRS (i.e. relative station elevation is the same as elevation until
topography is applied to the model). Once topography is applied,
relative station elevation will be set by
Data.project_stations_on_topography.

The other options are to leave relative elevation as 0.0 until
topography is applied, but this means the .dat file will not contain
station elevations unless there's topography. We can also set it to be
relative to grid centre point (i.e. highest station in the grid) to be
consistent with making rel_east and rel_north relative to the center.
This depends on feedback.