Add ability to get surface model from EDI data
To use, call `model_object.add_topography_from_data(data_object)` using
a Data object filled with EDI data.

This required some refactoring of the current topography methods.
`add_topography_to_model2` has been refactored to allow providing a
surface tuple containing ([lat], [lon], [elev]) to be used as an
elevation grid for interpolating the surface model.

A wrapper around the above function has been created,
`add_topography_from_data`, that will take a filled Data object
and use the station locations to generate the surface tuple, passing
it to `add_topography_to_model2`.

The rest of the workflow remains the same, including the
`Data.project_stations_on_topography` method placing the stations
below the surface model.

One thing to note is that in
`mesh_tools.interpolate_elevation_to_grid`, creating a numpy meshgrid
from the horizontal component of the station locations results in a
dense grid which gets flattned to arrays of length N^2 (where N is
number of stations). This causes a dimension mismatch between
horizontal and vertical dimensions which means interpolation can't
be performed. A workaround at the moment is to create a sparse grid
when not using the 'fast' method.