Add background image to phase tensor map
Users can now provide a geotiff to be displayed as the background image
of the phase tensor map. This is done by using 'imshow' to display a
window of the image that corresponds to the limits of the axes/basemap.

Users are responsible for selecting the bands to be displayed. If the
data is RGB/A or single band, then it's okay to leave the `bimg_band`
parameter as None (will get all available bands by default). However
if a specific band is required, it must be selected with the `bimg_band`
parameter.