Change plot pen depth 2D to take periods
Specifying period indicies would cause an exception if the period
index did not exist on all stations. It also causes issue becaue
the period for each index may not be the same across all stations.

The new approach is to specify a list of periods in seconds and a
tolerance. The nearest periods to the provided periods will be plotted.
If the nearest found period is outside of tolerance, then it will be
discarded and appear as a gap in the plot for that station.

Resolves #104