Fix ModEM data tests
Function EdiCollection.get_periods_by_stats iterates through station
periods, places them in a dict of {period: n_occurances} and orders
by n_occurances. A list is then generated from the dict (key, val) pairs
and this gets returned and used elsewhere, in this case in modem.Data.

In Python 2, dict ordering is not guaranteed, so after being placed
in the dictionary the periods are longer sorted. In Python 3, dict
ordering is by order of insertion, so the periods remain sorted. This
causes a mismatch between the ModEM data files written by Python 2 and
Python 3 as the periods are out of order.

The fix for this is to regenerate the expected data files using Python
3. If for whatever reason we need to pass these tests in future
using Python 2, the periods will need to be sorted in
EdiCollection.get_period_by_stats before being returned.