Fix bug with Occam2D EDI ordering
The order in which EDI files were read was affecting profile
generation. The order of east/north points used in linear regression
to get the profile line was dependent on EDI ordering. EDI ordering
was arbitrary due to the use of os.listdir to read the files, which
is affected by OS. The impact this had was that the Y-intercept of
the profile line could be different by a significant amount depending
on the platform MTPY is run on.

This change preserves the original behaviour. You can still provide
station_list and EDIs will be loaded in this order, or leave blank
to load using listdir - but with addition of a sort so behaviour is
consistent across platforms.

An addition is made that takes the optimize_path function from
modem.plot_slices and uses it to order the EDIs based on the
optimal path through the east/north points (essentially ordering
stations to get the 'smoothest' regression). This can be used by
providing `optimize_line=True' to the Occam2D data object before
generating the profile (note that `station_list` being present will
override it).